requirements.txt
pandas>=2.2
numpy>=1.26
scipy>=1.13
streamlit>=1.44
requests>=2.32
thefuzz>=0.22
python-dotenv>=1.0
